refactor: 로그인 api 함수 분리 #1
